Make APY daemon a library so we can fuzz it
The fuzzer binary needs access to some of the internal functions, so we
will make them public.
