+++ cli/listener/fuzz/.gitignore
@@ -1,0 +1,3 @@
+target
+corpus
+artifacts
