+++ cli/listener/fuzz/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "listener-fuzz"
+version = "0.0.0"
+authors = ["Automatically generated"]
+publish = false
+edition = "2018"
+
+[package.metadata]
+cargo-fuzz = true
+
+[dependencies]
+libfuzzer-sys = "0.4"
+arbitrary = "1.1.0"
+
+[dependencies.listener]
+path = ".."
+
+# Prevent this from interfering with workspaces
+[workspace]
+members = ["."]
+
+[[bin]]
+name = "apy_endpoint"
+path = "fuzz_targets/apy_endpoint.rs"
+test = false
+doc = false
